fixed the following
balance is not updated correctly both when
user buys something ( user inital amount 1000, after payment of 200 user balance was = 600)

when user logs out and logs back user balance is back to what they added, calculation is not updated in the folder.

when user adds more money, it is not incremanted what they already have, rather changed to that value.

added a statment to every method there is so if the user feels like they don't wish to continue they can go back to the previous menu